requirements.txt
colorama==0.4.5
hikari==2.0.0.dev110
hikari-lightbulb==2.2.4
